OSX/리눅스에서 컴파일되도록 수정함

+++ trunk/SipParser/SipUtility.cpp
@@ -117,5 +117,5 @@ void SipMakeCallIdName( char * pszCallId, int iCallIdSize )
 	iCallId = giCallId;
 	gclsMutex.release();
 
-	snprintf( pszCallId, iCallIdSize, "WCSS-%d-%d-%d", iCallId, sttTime.tv_sec, sttTime.tv_usec );
+	snprintf( pszCallId, iCallIdSize, "WCSS-%d-%d-%d", iCallId, (int)sttTime.tv_sec, (int)sttTime.tv_usec );
 }
